+++ modules/feature-comments.php
@@ -338,7 +338,15 @@ function delete_comment(&$comment_data, $target_id)
 	
 	for($i = 0; $i < $comment_count; $i++) {
 		if($comment_data[$i]->id == $target_id) {
-			$comment_data[$i]->message = "_[Deleted]_";
+			if(count($comment_data[$i]->replies) == 0) {
+				unset($comment_data[$i]);
+				// Reindex the comment list before returning
+				$comment_data = array_values($comment_data);
+			}
+			else {
+				unset($comment_data[$i]->username);
+				$comment_data[$i]->message = "_[Deleted]_";
+			}
 			return true;
 		}
 		if(count($comment_data[$i]->replies) > 0 &&
@@ -408,7 +416,7 @@ function render_comments($comments_data, $depth = 0)
 		$comment = $comments_data[$i];
 		
 		$result .= "\t<div class='comment' id='comment-$comment->id' data-comment-id='$comment->id'>\n";
-		$result .= "\t<p class='comment-header'><span class='name'>" . page_renderer::render_username($comment->username) . "</span> said:</p>";
+		$result .= "\t<p class='comment-header'><span class='name'>" . page_renderer::render_username($comment->username ?? "<em>Unknown</em>") . "</span> said:</p>";
 		$result .= "\t<div class='comment-body'>\n";
 		$result .= "\t\t" . parse_page_source($comment->message);
 		$result .= "\t</div>\n";